改造mybatis 抽出rdbms
