--- lodsve-mybatis/src/main/java/lodsve/mybatis/configs/annotations/EnableMyBatis.java
@@ -1,46 +1,0 @@
-/*
- * Copyright (C) 2018  Sun.Hao
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-package lodsve.mybatis.configs.annotations;
-
-import lodsve.core.configuration.EnableLodsve;
-import lodsve.mybatis.configs.MyBatisConfiguration;
-import lodsve.mybatis.datasource.DataSourceBeanDefinitionRegistrar;
-import lodsve.mybatis.datasource.DataSourceTransactionManagementConfiguration;
-import org.springframework.context.annotation.Import;
-
-import java.lang.annotation.*;
-
-/**
- * 通过注解的方式启动mybatis的配置.
- *
- * @author <a href="mailto:sunhao.java@gmail.com">sunhao(sunhao.java@gmail.com)</a>
- * @date 16/1/19 下午8:01
- */
-@Target(ElementType.TYPE)
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-@EnableLodsve
-@Import({MyBatisConfiguration.class, DataSourceTransactionManagementConfiguration.class, DataSourceBeanDefinitionRegistrar.class})
-public @interface EnableMyBatis {
-    /**
-     * 数据源名,多数据源,第一个数据源为默认值
-     *
-     * @return
-     */
-    String[] dataSource();
-}
