+++ lodsve-mybatis/src/main/java/lodsve/mybatis/dialect/AbstractDialect.java
@@ -21,8 +21,10 @@ import org.springframework.jdbc.datasource.DataSourceUtils;
 import org.springframework.util.Assert;
 
 import javax.sql.DataSource;
+import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
+import java.sql.SQLException;
 
 /**
  * 公用部分.
@@ -33,30 +35,38 @@ import java.sql.ResultSet;
 public abstract class AbstractDialect implements Dialect {
     @Override
     public String getCountSql(String sql) {
+        // TODO 优化
         StringBuilder sqlBuilder = new StringBuilder();
         return sqlBuilder.append("select count(*) from (").append(sql).append(") tmp_count").toString();
     }
 
     @Override
-    public boolean existTable(String tableName, DataSource dataSource) throws Exception {
+    public boolean existTable(String tableName, DataSource dataSource) throws SQLException {
         Assert.notNull(dataSource);
         Assert.hasText(tableName);
 
-        String name = dataSource.getConnection().getCatalog();
+
+        Connection connection = null;
         PreparedStatement ps = null;
         ResultSet resultSet = null;
 
         try {
-            ps = DataSourceUtils.doGetConnection(dataSource).prepareStatement(existTableSql(name, tableName));
+            connection = DataSourceUtils.getConnection(dataSource);
+
+            String name = connection.getCatalog();
+            ps = connection.prepareStatement(existTableSql(name, tableName));
             resultSet = ps.executeQuery();
 
             return resultSet.next() && resultSet.getInt("count") > 0;
         } finally {
+            if (resultSet != null) {
+                resultSet.close();
+            }
             if (ps != null) {
                 ps.close();
             }
-            if (resultSet != null) {
-                resultSet.close();
+            if (connection != null) {
+                DataSourceUtils.releaseConnection(connection, dataSource);
             }
         }
     }
@@ -64,9 +74,9 @@ public abstract class AbstractDialect implements Dialect {
     /**
      * 判断表是否存在的sql
      *
-     * @param schema
-     * @param tableName
-     * @return
+     * @param schema    schema
+     * @param tableName table Name
+     * @return 判断表是否存在的sql
      */
     abstract String existTableSql(String schema, String tableName);
 }
