--- lodsve-mybatis/src/main/java/lodsve/mybatis/properties/DruidProperties.java
@@ -1,111 +1,0 @@
-/*
- * Copyright (C) 2018  Sun.Hao
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-package lodsve.mybatis.properties;
-
-import lodsve.core.properties.relaxedbind.annotations.ConfigurationProperties;
-
-/**
- * druid配置.
- *
- * @author <a href="mailto:sunhao.java@gmail.com">sunhao(sunhao.java@gmail.com)</a>
- * @date 2018-2-8-0008 15:05
- */
-@ConfigurationProperties(prefix = "lodsve.druid", locations = "${params.root}/framework/druid.properties")
-public class DruidProperties {
-    private String filters = "stat\\,wall";
-    private boolean enableMonitor = false;
-    private String path = "/druid/*";
-    private String resetEnable = "true";
-    private String user;
-    private String password;
-    private String allow;
-    private String deny;
-    private String remoteAddress;
-
-    public String getFilters() {
-        return filters;
-    }
-
-    public void setFilters(String filters) {
-        this.filters = filters;
-    }
-
-    public boolean isEnableMonitor() {
-        return enableMonitor;
-    }
-
-    public void setEnableMonitor(boolean enableMonitor) {
-        this.enableMonitor = enableMonitor;
-    }
-
-    public String getPath() {
-        return path;
-    }
-
-    public void setPath(String path) {
-        this.path = path;
-    }
-
-    public String getResetEnable() {
-        return resetEnable;
-    }
-
-    public void setResetEnable(String resetEnable) {
-        this.resetEnable = resetEnable;
-    }
-
-    public String getUser() {
-        return user;
-    }
-
-    public void setUser(String user) {
-        this.user = user;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getAllow() {
-        return allow;
-    }
-
-    public void setAllow(String allow) {
-        this.allow = allow;
-    }
-
-    public String getDeny() {
-        return deny;
-    }
-
-    public void setDeny(String deny) {
-        this.deny = deny;
-    }
-
-    public String getRemoteAddress() {
-        return remoteAddress;
-    }
-
-    public void setRemoteAddress(String remoteAddress) {
-        this.remoteAddress = remoteAddress;
-    }
-}
