+++ lodsve-rdbms/pom.xml
@@ -1,0 +1,66 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <!-- 配置父 -->
+    <parent>
+        <groupId>com.lodsve</groupId>
+        <artifactId>lodsve-framework</artifactId>
+        <version>2.6.8-SNAPSHOT</version>
+        <relativePath>../pom.xml</relativePath>
+    </parent>
+
+    <artifactId>lodsve-rdbms</artifactId>
+    <packaging>jar</packaging>
+    <name>lodsve rdbms</name>
+    <description>关系型数据库数据源</description>
+    <url>https://github.com/lodsve/lodsve-framework</url>
+
+    <dependencies>
+        <!-- required dependency -->
+        <dependency>
+            <groupId>com.google.guava</groupId>
+            <artifactId>guava</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-jdbc</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.aspectj</groupId>
+            <artifactId>aspectjweaver</artifactId>
+        </dependency>
+
+        <!-- optional dependency -->
+        <dependency>
+            <groupId>com.alibaba</groupId>
+            <artifactId>druid</artifactId>
+            <optional>true</optional>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-web</artifactId>
+            <optional>true</optional>
+        </dependency>
+        <dependency>
+            <groupId>javax.servlet</groupId>
+            <artifactId>javax.servlet-api</artifactId>
+            <optional>true</optional>
+        </dependency>
+        <dependency>
+            <groupId>org.flywaydb</groupId>
+            <artifactId>flyway-core</artifactId>
+            <optional>true</optional>
+        </dependency>
+
+        <!-- lodsve -->
+        <dependency>
+            <groupId>com.lodsve</groupId>
+            <artifactId>lodsve-core</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.lodsve</groupId>
+            <artifactId>lodsve-3rd</artifactId>
+        </dependency>
+    </dependencies>
+</project>
