+++ lodsve-rdbms/src/main/java/lodsve/rdbms/Constants.java
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package lodsve.mybatis.utils;
+package lodsve.rdbms;
 
 /**
  * 常量.
@@ -27,8 +27,12 @@ public class Constants {
     private Constants() {
     }
 
+    public static final String DATA_SOURCE_ATTRIBUTE_NAME = "dataSource";
+
     public static final String DATA_SOURCE_BEAN_NAME = "lodsveDataSource";
-    public static final String MYBATIS_SQL_SESSION_FACTORY_BANE_NAME = "sqlSessionFactory";
-    public static final String MAPPER_SCANNER_CONFIGURER_BANE_NAME = "mapperScannerConfigurer";
-    public static final String ID_GENERATOR_BANE_NAME = "idGenerator";
+    public static final String REAL_DATA_SOURCE_BEAN_NAME = "lodsveRealDataSource";
+
+    public static final String DRUID_DATA_SOURCE_CLASS = "com.alibaba.druid.pool.DruidDataSource";
+
+    public static final String FLYWAY_BEAN_NAME = "lodsveFlyway";
 }
