+++ lodsve-rdbms/src/main/java/lodsve/rdbms/dynamic/DynamicDataSourceAspect.java
@@ -1,0 +1,55 @@
+/*
+ * Copyright (C) 2018  Sun.Hao
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package lodsve.rdbms.dynamic;
+
+import lodsve.core.utils.StringUtils;
+import lodsve.rdbms.annotations.SwitchDataSource;
+import org.aspectj.lang.ProceedingJoinPoint;
+import org.aspectj.lang.annotation.Around;
+import org.aspectj.lang.annotation.Aspect;
+import org.aspectj.lang.reflect.MethodSignature;
+import org.springframework.stereotype.Component;
+
+import java.lang.reflect.Method;
+
+/**
+ * aop动态切换数据源.
+ *
+ * @author <a href="mailto:sunhao.java@gmail.com">sunhao(sunhao.java@gmail.com)</a>
+ * @date 2017/12/14 下午6:16
+ */
+@Aspect
+@Component
+public class DynamicDataSourceAspect {
+    @Around("@annotation(lodsve.rdbms.annotations.SwitchDataSource)")
+    public Object around(ProceedingJoinPoint point) throws Throwable {
+        String dataSource = StringUtils.EMPTY;
+        MethodSignature methodSignature = (MethodSignature) point.getSignature();
+        Method method = methodSignature.getMethod();
+        if (method.isAnnotationPresent(SwitchDataSource.class)) {
+            SwitchDataSource annotation = method.getAnnotation(SwitchDataSource.class);
+            // 取出注解中的数据源名
+            dataSource = annotation.value();
+        }
+
+        try (DataSourceHolder dsh = DataSourceHolder.getInstance()) {
+            dsh.set(dataSource);
+            return point.proceed();
+        }
+    }
+}
